Working build, just needs animations and polish
