+++ src/main/java/io/jenkins/plugins/designlibrary/Root.java
@@ -3,18 +3,12 @@ package io.jenkins.plugins.designlibrary;
 import hudson.Extension;
 import hudson.model.RootAction;
 import hudson.util.HttpResponses;
-import jenkins.model.Jenkins;
 import net.sf.json.JSONArray;
-import org.apache.commons.jelly.JellyException;
-import org.jenkins.ui.icon.IconSet;
 import org.kohsuke.stapler.HttpResponse;
-import org.kohsuke.stapler.Stapler;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
 
-import java.io.IOException;
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * Entry point to all the UI samples.
@@ -44,76 +38,9 @@ public class Root implements RootAction {
         return null;
     }
 
-    public static class SearchResult {
-        private String icon;
-        private String title;
-        private String url;
-        private List<SearchResult> children;
-
-        public SearchResult() {
-        }
-
-        public SearchResult(String title, String url) {
-            this.title = title;
-            this.url = url;
-        }
-
-        public String getIcon() {
-            return icon;
-        }
-
-        public void setIcon(String icon) {
-            this.icon = icon;
-        }
-
-        public String getTitle() {
-            return title;
-        }
-
-        public void setTitle(String title) {
-            this.title = title;
-        }
-
-        public String getUrl() {
-            return url;
-        }
-
-        public void setUrl(String url) {
-            this.url = url;
-        }
-
-        public List<SearchResult> getChildren() {
-            return children;
-        }
-
-        public void setChildren(List<SearchResult> children) {
-            this.children = children;
-        }
-    }
-
-    public HttpResponse doSearch(StaplerRequest req, StaplerResponse response) {
-        final StaplerRequest request = Stapler.getCurrentRequest();
+    public HttpResponse doSearch(StaplerRequest request, StaplerResponse response) {
         final String query = request.getParameter("query");
-
-        List<SearchResult> searchResults = getAll().stream()
-                .filter(sample -> sample.getDisplayName().toLowerCase().contains(query.toLowerCase()))
-                .limit(5)
-                .map(sample -> {
-            SearchResult searchResult = new SearchResult();
-            searchResult.setIcon(IconSet.getSymbol(sample.getIconFileName().substring(7),
-                    "", "", "", "design-library", ""));
-            searchResult.setTitle(sample.getDisplayName());
-            searchResult.setUrl(Jenkins.get().getRootUrl() + "design-library/" + sample.getUrlName());
-                    try {
-                        searchResult.setChildren(sample.getHeadings(request, response));
-                    } catch (JellyException | IOException e) {
-                        throw new RuntimeException(e);
-                    }
-                    return searchResult;
-        })
-                .collect(Collectors.toList());
-
-        return HttpResponses.okJSON(JSONArray.fromObject(searchResults));
+        return HttpResponses.okJSON(JSONArray.fromObject(SearchHelper.getSearchResults(query, request, response)));
     }
 
     public List<UISample> getAll() {
